restarea-service/requirements.txt
asyncio==4.0.0
asyncio-mqtt==0.16.2
dotenv==0.9.9
paho-mqtt==2.1.0
python-dotenv==1.2.1
